ref: Move test folder and disable mirror.test for general usage;
